Fix integer overflow for expected staking time.

+++ src/qt/bitcoingui.cpp
@@ -890,7 +890,7 @@ void BitcoinGUI::updateStakingIcon()
     if (nLastCoinStakeSearchInterval && nAverageWeight)
     {
         uint64 nNetworkWeight = GetPoSVKernelPS();
-        unsigned nEstimateTime = nTargetSpacing * nNetworkWeight / nTotalWeight;
+        uint64 nEstimateTime = nTargetSpacing * nNetworkWeight / nTotalWeight;
 
         QString text;
         if (nEstimateTime < 60)
